requirements.txt
requests==2.32.3
aiohttp==3.10.5
beautifulsoup4==4.12.3
huggingface-hub==0.25.1
tenacity==9.0.0
loguru==0.7.2
python-dotenv==1.0.1
polars==1.9.0
duckdb==1.1.1
